:sparkles: Refactored Code Structure
