+++ package.json
@@ -3,8 +3,10 @@
   "version": "0.1.2",
   "private": true,
   "dependencies": {
+    "bootstrap": "^3.3.7",
     "instascan": "^1.0.0",
     "jquery": "^3.2.1",
+    "qrcode.react": "^0.7.2",
     "react": "^16.0.0",
     "react-bootstrap": "^0.31.3",
     "react-dom": "^16.0.0",
