--- src/App.js
@@ -1,22 +1,0 @@
-import React, { Component } from 'react';
-import logo from './logo.svg';
-import './App.css'; 
-
-
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h1 className="App-title">Why join Navy?, YOU JOIN NAVY!!</h1>
-        </header>
-        <p className="App-intro">
-          HI</p>
-        
-      </div>
-    );
-  }
-}
-
-export default App;
