+++ src/components/App/index.js
@@ -1,0 +1,16 @@
+import React from 'react'
+
+import logo from '../assets/logo.svg'
+import './App.css'
+
+const App = () => (
+  <div className="App">
+    <header className="App-header">
+      <img src={logo} className="App-logo" alt="logo" />
+      <h1 className="App-title">Why join navy when you can be a pirate?</h1>
+    </header>
+    <p className="App-intro">Hi!</p>
+  </div>
+)
+
+export default App
