+++ src/components/Footer.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default () => (
+  <footer className="footer">
+    <div className="container">
+      <span className="text-muted">copynotright &copy; StupidHackathonTH.</span>
+    </div>
+  </footer>
+)
