--- src/index.css
@@ -1,18 +1,0 @@
-body {
-  margin: 0;
-  padding: 0;
-  font-family: sans-serif;
-  background-color: #1a1d23 !important;
-}
-.w100p{
-	width: 100%;
-}
-.bg-w{
-	padding-top: 5px;
-	margin: 5px;
-	background-color: #fff !important;
-}
-.tx-w{
-	color: #636363 !important;
-	font-size: 5em !important;
-}
